+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fixed possibility of customizations via `UIAppearance` being overridden.
 * Fixed issue related to sending a push notification with `(null)` text. 
 * Fixed typo for property `conversationLabelColor` on `LYRUIConversationTableViewCell`.
+* Fixed typo in C function signature `LYRUILightGrayColor()`.
 
 ### Enhancements
 * Refactored internal constants.