+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 ### Public API Changes
 * Added `bubbleViewCornerRadius` property to `LYRUIMessageCollectionViewCell`.  
 * Added `avatarImageViewCornerRadius` property to `LYRUIMessageCollectionViewCell`.
+* Changed `backgroundColor` property to `cellBackgroundColor` on `LYRUIConversationTableViewCell`.
 
 ### Bug Fixes
 * Fixed possibility of customizations via `UIAppearance` being overridden.
-* Fixed issue related to sending a push notification with `(null)` text. 
+* Fixed issue related to sending a push notification with `(null)` text.
 * Fixed typo for property `conversationLabelColor` on `LYRUIConversationTableViewCell`.
 * Fixed typo in C function signature `LYRUILightGrayColor()`.